+++ spring-web-support/src/main/java/io/springperf/web/support/arg/provider/HttpServletResponseProvider.java
@@ -4,17 +4,14 @@ import io.springperf.web.context.WebContext;
 import io.springperf.web.core.arg.StaticArgumentResolver;
 import io.springperf.web.core.arg.provider.StaticArgumentResolverProvider;
 import io.springperf.web.core.mapping.MappingHandlerMethod;
+import io.springperf.web.support.servlet.ServletAttribute;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.core.MethodParameter;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
 
 public class HttpServletResponseProvider implements StaticArgumentResolverProvider {
 
-    private final StaticArgumentResolver resolver = (request, response) -> {
-        ServletRequestAttributes attributes = (ServletRequestAttributes) RequestContextHolder.getRequestAttributes();
-        return attributes.getResponse();
-    };
+    private final StaticArgumentResolver resolver = (request, response) ->
+            ServletAttribute.getAdapterContext(request, response).getResponse();
 
     @Override
     public boolean supports(MethodParameter parameter, MappingHandlerMethod mappingContext) {