+++ spring-web-support/src/main/java/io/springperf/web/support/arg/provider/WebRequestArgumentResolverProvider.java
@@ -4,36 +4,20 @@ import io.springperf.web.context.WebContext;
 import io.springperf.web.core.arg.StaticArgumentResolver;
 import io.springperf.web.core.arg.provider.StaticArgumentResolverProvider;
 import io.springperf.web.core.mapping.MappingHandlerMethod;
+import io.springperf.web.support.servlet.ServletAttribute;
+import io.springperf.web.support.servlet.context.ServletAdapterContext;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.core.MethodParameter;
-import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.context.request.WebRequest;
 
 public class WebRequestArgumentResolverProvider implements StaticArgumentResolverProvider {
 
     private final StaticArgumentResolver resolver = (request, response) -> {
-        ServletWebRequest servletWebRequest = buildServletWebRequest();
-        return servletWebRequest;
+        ServletAdapterContext adapterContext = ServletAttribute.getAdapterContext(request, response);
+        return new ServletWebRequest(adapterContext.getRequest(), adapterContext.getResponse());
     };
-
-    protected ServletWebRequest buildServletWebRequest() {
-        ServletWebRequest servletWebRequest = null;
-        Object attrs = RequestContextHolder.getRequestAttributes();
-        if (attrs instanceof org.springframework.web.context.request.ServletRequestAttributes) {
-            org.springframework.web.context.request.ServletRequestAttributes servletRequestAttributes =
-                    (org.springframework.web.context.request.ServletRequestAttributes) attrs;
-            HttpServletRequest httpReq = servletRequestAttributes.getRequest();
-            HttpServletResponse httpResp = servletRequestAttributes.getResponse();
-            if (httpResp != null) {
-                servletWebRequest = new ServletWebRequest(httpReq, httpResp);
-            } else {
-                servletWebRequest = new ServletWebRequest(httpReq);
-            }
-        }
-        return servletWebRequest;
-    }
 
     @Override
     public boolean supports(MethodParameter parameter, MappingHandlerMethod mappingContext) {